Add the ability to share arbitrary file by AirDrop
I added the ability to share an arbitrary file.  I've only tested it via AirDrop because I don't use any other sharing services but I needed this to be able to start sharing a file from the command line.  I hope that this meets your goals and requirements and that this is a helpful contribution to your project.  Thank you!